# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/categories": {
      "post": {
        "description": "Create a custom transaction category. Subject to plan limits. Supports 2-level hierarchy (parent → child).",
        "operationId": "CategoriesController_create_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Category created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation error, max depth exceeded, or parent type mismatch"
          },
          "403": {
            "description": "Category limit reached or system category modification"
          },
          "409": {
            "description": "Category name already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create custom category",
        "tags": [
          "Categories"
        ]
      },
      "get": {
        "description": "Get all categories for the authenticated workspace. Includes system categories (shared) and workspace custom categories. Returns tree structure by default.",
        "operationId": "CategoriesController_findAll_v1",
        "parameters": [
          {
            "name": "type",
            "required": false,
            "in": "query",
            "description": "Filter by transaction type",
            "schema": {
              "example": "EXPENSE",
              "type": "string",
              "enum": [
                "EXPENSE",
                "INCOME",
                "TRANSFER"
              ]
            }
          },
          {
            "name": "includeInactive",
            "required": false,
            "in": "query",
            "description": "Include inactive categories (default: false)",
            "schema": {
              "default": false,
              "example": false,
              "type": "boolean"
            }
          },
          {
            "name": "flat",
            "required": false,
            "in": "query",
            "description": "Return flat list instead of tree structure (default: false)",
            "schema": {
              "default": false,
              "example": false,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category tree (or flat list if ?flat=true)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryTreeResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List categories",
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/v1/categories/{id}": {
      "get": {
        "description": "Get details of a specific category including its subcategories.",
        "operationId": "CategoriesController_findById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category ID (CUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category details with children",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryWithChildrenDto"
                }
              }
            }
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get category",
        "tags": [
          "Categories"
        ]
      },
      "put": {
        "description": "Update a custom category. System categories cannot be modified. Type and parent cannot be changed.",
        "operationId": "CategoriesController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category ID (CUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Category updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Cannot modify system categories"
          },
          "404": {
            "description": "Category not found"
          },
          "409": {
            "description": "Category name already exists"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update category",
        "tags": [
          "Categories"
        ]
      },
      "delete": {
        "description": "Permanently delete a custom category. System categories cannot be deleted. Categories with transactions can only be deactivated.",
        "operationId": "CategoriesController_delete_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category ID (CUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Category deleted successfully"
          },
          "403": {
            "description": "Cannot delete system categories"
          },
          "404": {
            "description": "Category not found"
          },
          "409": {
            "description": "Category has transactions, deactivate instead"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete category",
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/v1/categories/{id}/deactivate": {
      "patch": {
        "description": "Deactivate a custom category to hide it from selection lists while preserving existing transaction references.",
        "operationId": "CategoriesController_deactivate_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category ID (CUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category deactivated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Cannot deactivate system categories"
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate category",
        "tags": [
          "Categories"
        ]
      }
    },
    "/api/v1/categories/{id}/reactivate": {
      "patch": {
        "description": "Reactivate a previously deactivated custom category.",
        "operationId": "CategoriesController_reactivate_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Category ID (CUID)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Category reactivated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CategoryResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Cannot modify system categories"
          },
          "404": {
            "description": "Category not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Reactivate category",
        "tags": [
          "Categories"
        ]
      }
    },
...
  },
  "info": {
    "title": "Facets API",
    "description": "Professional Finance Tracker SaaS API",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Category name",
            "example": "Groceries",
            "maxLength": 100
          },
          "type": {
            "type": "string",
            "description": "Transaction type this category applies to",
            "enum": [
              "EXPENSE",
              "INCOME",
              "TRANSFER"
            ],
            "example": "EXPENSE"
          },
          "parentId": {
            "type": "string",
            "description": "Parent category ID. If set, this becomes a subcategory. Max 2-level depth.",
            "example": "cm3xk7z9w0001jn08abc12345"
          },
          "icon": {
            "type": "string",
            "description": "Icon identifier for the app UI",
            "example": "shopping-cart"
          },
          "color": {
            "type": "string",
            "description": "Hex color code for UI",
            "example": "#FF6B6B"
          },
          "sortOrder": {
            "type": "number",
            "description": "Display order (lower = first)",
            "example": 0,
            "default": 0
          }
        },
        "required": [
          "name",
          "type"
        ]
      },
      "CategoryResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm3xk7z9w0001jn08abc12345"
          },
          "name": {
            "type": "string",
            "example": "Food & Drinks"
          },
          "type": {
            "type": "string",
            "enum": [
              "EXPENSE",
              "INCOME",
              "TRANSFER"
            ],
            "example": "EXPENSE"
          },
          "parentId": {
            "type": "string",
            "example": "cm3xk7z9w0001jn08abc12345"
          },
          "icon": {
            "type": "string",
            "example": "utensils"
          },
          "color": {
            "type": "string",
            "example": "#FF6B6B"
          },
          "isSystem": {
            "type": "boolean",
            "example": false,
            "description": "System categories cannot be edited or deleted"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "sortOrder": {
            "type": "number",
            "example": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "isSystem",
          "isActive",
          "sortOrder",
          "createdAt",
          "updatedAt"
        ]
      },
      "CategoryWithChildrenDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cm3xk7z9w0001jn08abc12345"
          },
          "name": {
            "type": "string",
            "example": "Food & Drinks"
          },
          "type": {
            "type": "string",
            "enum": [
              "EXPENSE",
              "INCOME",
              "TRANSFER"
            ],
            "example": "EXPENSE"
          },
          "parentId": {
            "type": "string",
            "example": "cm3xk7z9w0001jn08abc12345"
          },
          "icon": {
            "type": "string",
            "example": "utensils"
          },
          "color": {
            "type": "string",
            "example": "#FF6B6B"
          },
          "isSystem": {
            "type": "boolean",
            "example": false,
            "description": "System categories cannot be edited or deleted"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "sortOrder": {
            "type": "number",
            "example": 0
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "example": "2025-01-15T10:30:00.000Z"
          },
          "children": {
            "description": "Subcategories (second level)",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryResponseDto"
            }
          }
        },
        "required": [
          "id",
          "name",
          "type",
          "isSystem",
          "isActive",
          "sortOrder",
          "createdAt",
          "updatedAt",
          "children"
        ]
      },
      "CategoryTreeResponseDto": {
        "type": "object",
        "properties": {
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryWithChildrenDto"
            }
          },
          "total": {
            "type": "number",
            "example": 15,
            "description": "Total number of categories (including children)"
          }
        },
        "required": [
          "categories",
          "total"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Category name",
            "example": "Groceries",
            "maxLength": 100
          },
          "icon": {
            "type": "string",
            "description": "Icon identifier for the app UI",
            "example": "shopping-cart"
          },
          "color": {
            "type": "string",
            "description": "Hex color code for UI",
            "example": "#FF6B6B"
          },
          "sortOrder": {
            "type": "number",
            "description": "Display order (lower = first)",
            "example": 0,
            "default": 0
          }
        }
      },
...
    }
  }
}